Split tests into smaller modules
And rename them so they're detected by pytest.
